+++ src/components/charts/discoverChartDetail/constants.ts
@@ -1,41 +1,7 @@
-import { RouteComponentProps } from 'react-router'
 import { ReactComponent as GotToBuildDeploy } from '../../../assets/icons/go-to-buildanddeploy@2x.svg'
 import { ReactComponent as GoToEnvOverride } from '../../../assets/icons/go-to-envoverride@2x.svg'
 import { DOCUMENTATION } from '../../../config'
-import { ChartValuesType } from '../charts.types'
-
-export interface DiscoverChartDetailsProps extends RouteComponentProps<{ chartId: string }> {}
-
-export interface DeploymentProps {
-    icon?: string
-    chartName?: string
-    name?: string
-    chartId: string
-    appStoreApplicationName?: string
-    deprecated: boolean
-    isGitOpsConfigAvailable
-    showGitOpsWarningModal
-    toggleGitOpsWarningModal
-    availableVersions: Map<number, { id; version }>
-}
-
-export interface PrimaryOptionType {
-    icon: React.FunctionComponent<any>
-    title: string
-    subtitle: string
-    valueType: string
-    noDataSubtitle?: string[]
-    helpLink?: string
-}
-export interface ChartVersionSelectorModalType {
-    closePopup: () => void
-    appStoreApplicationName: string
-    appIconUrl: string
-    onError: (e) => void
-    handleDeploy: () => void
-    deployedChartValueList: ChartValuesType[]
-    presetChartValueList: ChartValuesType[]
-}
+import { PrimaryOptionType } from './types'
 
 export const ValueType = {
     PRESET: 'preset',
